api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
slowapi==0.1.9
sqlmodel==0.0.22
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
fastapi-mail==1.4.1
apscheduler==3.10.4
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0
Pillow==11.0.0
jinja2==3.1.4
openpyxl==3.1.5
